requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20241021
distlib==0.4.0
Django==5.2.3
django-import-export==4.3.9
et_xmlfile==2.0.0
filelock==3.19.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
odfpy==1.4.1
openpyxl==3.1.5
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tablib==3.8.0
tabulate==0.9.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.34.0
websockets==15.0.1
Werkzeug==3.1.3
xlrd==1.2.0
xlwt==1.3.0
